+++ include/vml/gen_vector_vml.sh
@@ -148,7 +148,7 @@ echo "/**@}*/"
 
 
 ## 2arg math
-math=(sin sqrt sinh asin asinh tan tanh atan atanh ceil floor sign)
+math=(sin sqrt sinh asin asinh tan tanh atan atanh ceil floor sign exp)
 for math in ${math[@]}; do
 echo "
 /**